Settle pnl: Move fee-computation to PerpMarket
So it can be used in client code to decide what accounts to settle.
